Added creating a shortcut called "Puppet Report GUI" to the 'Start Menu'  This functionality only works in .Net Framework 4.0 or above for the moment.

+++ build/chocolateyUninstall.ps1
@@ -4,8 +4,20 @@ try {
   $installDir = Join-Path (Join-Path $PSHome "Modules") "POSHPuppetReports"
 
   # Remove the folder from the powershell modules directory
-  [void] (Remove-Item $installDir -Recurse -Confirm:$false -Force -ErrorAction 'Stop')
-     
+  if (Test-Path  $installDir) {
+    [void] (Remove-Item $installDir -Recurse -Confirm:$false -Force -ErrorAction 'Stop')
+  }
+  else
+  {
+    Write-Debug "Installation directory doesn't exist"
+  }
+
+  # Removing shortcuts for the reports into the "Start Menu"
+  $shortcutPath = ([Environment]::GetFolderPath('CommonStartMenu')) + '\POSH Puppet Reports\Report GUI.lnk'
+  if (Test-Path $shortcutPath) {
+    [void] (Remove-Item $shortcutPath -Recurse -Confirm:$false -Force -ErrorAction 'Stop')
+  }
+
   Write-ChocolateySuccess "$packageName"
 } catch {
   Write-ChocolateyFailure "$packageName" "$($_.Exception.Message)"
